forkify: added config file
