+++ package.json
@@ -25,6 +25,7 @@
     "babel-polyfill": "^6.26.0",
     "core-js": "^3.16.3",
     "core.js": "^0.4.2",
+    "fractional": "^1.0.0",
     "regenerator-runtime": "^0.13.9"
   }
 }
