+++ src/js/config.js
@@ -1,0 +1,5 @@
+// this file will contain all the varibles which should be constant and reused through out this project
+
+// so if want to change some configurations we do it using the config
+
+export const API_URL = 'https://forkify-api.herokuapp.com/api/v2/recipes';
