+++ src/js/rough.js
@@ -1,0 +1,178 @@
+import 'babel-polyfill';
+// import 'core-js/stable';
+// import 'regenerator-runtime/runtime'
+import icons from '../img/icons.svg'; // parcel 1
+// import icons from 'url: ../img/icons.svg'; // parcel 2
+
+const recipeContainer = document.querySelector('.recipe');
+
+// This is promisifyin a setTimeout callback function
+const timeout = function (s) {
+  return new Promise(function (_, reject) {
+    setTimeout(function () {
+      reject(new Error(`Request took too long! Timeout after ${s} second`));
+    }, s * 1000);
+  });
+};
+
+// https://forkify-api.herokuapp.com/v2
+
+///////////////////////////////////////
+
+// creating a generic function for the spinner
+const renderSpinner = function (parentEl) {
+  const markup = `
+  <div class="spinner">
+    <svg>
+      <use href="${icons}#icon-loader"></use>
+    </svg>
+  </div>
+  `
+  parentEl.innerHTML = ''; //clearing parent html first
+  parentEl.insertAdjacentHTML('afterbegin', markup)
+};
+
+// creating an async function that will fetch our recipe the API
+const showRecipe = async function () {
+  try {//success
+    const id = window.location.hash.slice(1); //? window.location refers to the whole url, this how you get the hash
+    if(!id) return;
+  
+    // TODO render spinner
+    renderSpinner(recipeContainer);
+    // TODO 1) loading recipe
+    const response = await fetch(
+      `https://forkify-api.herokuapp.com/api/v2/recipes/${id}`
+    ); //* making an ajax call to the API
+    const data = await response.json();
+    if (!response.ok) throw new Error(`${data.message} (${response.status})`); //catching error in the fetch
+
+    let { recipe } = data.data;
+    recipe = {
+      id: recipe.id,
+      title: recipe.title,
+      publisher: recipe.publisher,
+      sourceUrl: recipe.source_url,
+      image: recipe.image_url,
+      servings: recipe.servings,
+      cookingTime: recipe.cooking_time,
+      ingredients: recipe.ingredients,
+    }; //formatting recipe object keys
+
+    // console.log(recipe);
+
+    // TODO 2) rendering recipe
+
+    const markup = `
+    <figure class="recipe__fig">
+          <img src="${recipe.image}" alt="${
+      recipe.title
+    }" class="recipe__img" />
+          <h1 class="recipe__title">
+            <span>${recipe.title}</span>
+          </h1>
+        </figure>
+
+        <div class="recipe__details">
+          <div class="recipe__info">
+            <svg class="recipe__info-icon">
+              <use href="${icons}#icon-clock"></use>
+            </svg>
+            <span class="recipe__info-data recipe__info-data--minutes">${
+              recipe.cookingTime
+            }</span>
+            <span class="recipe__info-text">minutes</span>
+          </div>
+          <div class="recipe__info">
+            <svg class="recipe__info-icon">
+              <use href="${icons}#icon-users"></use>
+            </svg>
+            <span class="recipe__info-data recipe__info-data--people">${
+              recipe.servings
+            }</span>
+            <span class="recipe__info-text">servings</span>
+
+            <div class="recipe__info-buttons">
+              <button class="btn--tiny btn--increase-servings">
+                <svg>
+                  <use href="${icons}#icon-minus-circle"></use>
+                </svg>
+              </button>
+              <button class="btn--tiny btn--increase-servings">
+                <svg>
+                  <use href="${icons}#icon-plus-circle"></use>
+                </svg>
+              </button>
+            </div>
+          </div>
+
+          <div class="recipe__user-generated">
+            <svg>
+              <use href="${icons}#icon-user"></use>
+            </svg>
+          </div>
+          <button class="btn--round">
+            <svg class="">
+              <use href="${icons}#icon-bookmark-fill"></use>
+            </svg>
+          </button>
+        </div>
+
+        <div class="recipe__ingredients">
+          <h2 class="heading--2">Recipe ingredients</h2>
+          <ul class="recipe__ingredient-list">
+          ${recipe.ingredients
+            .map(ing => {
+              return `
+              <li class="recipe__ingredient">
+              <svg class="recipe__icon">
+                <use href="${icons}#icon-check"></use>
+              </svg>
+              <div class="recipe__quantity">${ing.quantity}</div>
+              <div class="recipe__description">
+                <span class="recipe__unit">${ing.unit}</span>
+                ${ing.description}
+              </div>
+            </li>
+            `;
+            })
+            .join('')}
+          </ul>
+        </div>
+
+        <div class="recipe__directions">
+          <h2 class="heading--2">How to cook it</h2>
+          <p class="recipe__directions-text">
+            This recipe was carefully designed and tested by
+            <span class="recipe__publisher">${
+              recipe.publisher
+            }</span>. Please check out
+            directions at their website.
+          </p>
+          <a
+            class="btn--small recipe__btn"
+            href="${recipe.sourceUrl}"
+            target="_blank"
+          >
+            <span>Directions</span>
+            <svg class="search__icon">
+              <use href="${icons}#icon-arrow-right"></use>
+            </svg>
+          </a>
+        </div>
+    `;
+
+    recipeContainer.innerHTML = ''; // removing the message
+    recipeContainer.insertAdjacentHTML('afterbegin', markup); // inserting the markup into the HTML
+  } catch (err) {
+    //error
+    alert(err);
+  }
+};
+
+showRecipe();
+
+['hashchange', 'load'].forEach(event => window.addEventListener(event, showRecipe)) // to do eventlisteners for the same event
+
+
+// window.addEventListener('hashchange', showRecipe) //listening for hash change event
